+++ src/theme.rs
@@ -24,6 +24,14 @@ pub enum BinaryColorTheme {
 
     /// An on/off OLED-like display with a dark blue background and light blue pixels
     OledBlue,
+
+    /// Custom binary color theme/mapping
+    Custom {
+        /// The color used for the "off" state pixels.
+        color_off: Rgb888,
+        /// The color used for the "on" state pixels.
+        color_on: Rgb888,
+    },
 }
 
 fn map_color(color: Rgb888, color_off: Rgb888, color_on: Rgb888) -> Rgb888 {
@@ -38,6 +46,10 @@ impl BinaryColorTheme {
     pub(crate) fn convert(self, color: Rgb888) -> Rgb888 {
         match self {
             BinaryColorTheme::Default => color,
+            BinaryColorTheme::Custom {
+                color_off,
+                color_on,
+            } => map_color(color, color_off, color_on),
             BinaryColorTheme::Inverted => {
                 Rgb888::new(255 - color.r(), 255 - color.g(), 255 - color.b())
             }